modified ATH adjustment code to use tot_ener returned by psy model and not the maximum of the PCM sample values.
We need to replace this with JD's formula.   We need to decide
if this is really worth while.  My preference is to take the
most conservative ATH possible, but this seems to produce too
high bitrates for VBR modes. (this may no longer be true
not that masings are computed in all scalefactor bands)


git-svn-id: svn://svn.code.sf.net/p/lame/svn/trunk/lame@1983 981c7233-bbb0-e711-95a9-001517a2e1a4