Put in fix for resampling bug in fill_buffer_resample.
Code assumed it would be fed at least BLACKSIZE samples.
For each input buffer, the code would save the last BLACKSIZE
samples for use for the next call to fill_buffer_resample.

If the code recieves less than blacksize samples, it will
now shift in just as many new samples as are available into
the save buffer.

Not tested :-(


git-svn-id: svn://svn.code.sf.net/p/lame/svn/trunk/lame@2007 981c7233-bbb0-e711-95a9-001517a2e1a4