Make autotools files compatible with automake >= 1.12
Newer autotools versions don't seem to support ansi2knr, which is probably
not a problem, given the last decades of compilers accepting ANSI C (and the
rest of the LAME code being written in ANSI C itself).

This fixes ticket #60, with a patch of Brennan Shacklett, from the Gentoo
community.


git-svn-id: svn://svn.code.sf.net/p/lame/svn/trunk/lame@6309 981c7233-bbb0-e711-95a9-001517a2e1a4