Replace mpeg decoder with external libmpg123 - step 1/2 part 1/3
Step 1 is to allow to use external libmp123 (preferred) or
interal mpglib to decode mpeg audio.

Patch by Thomas Orgis <thomas-forum@orgis.org>, with some
minor modifications to the autotools build by me.

Part 1: commit the source files and the input files for the build.



Step 2 would be to remove the internal mpglib.



git-svn-id: svn://svn.code.sf.net/p/lame/svn/trunk/lame@6463 981c7233-bbb0-e711-95a9-001517a2e1a4